Add unstakeAllFor unit tests
